# Track each SST's timestamp information as user properties