--- a/db/db_with_timestamp_basic_test.cc
+++ b/db/db_with_timestamp_basic_test.cc
@@ -619,12 +619,12 @@ TEST_F(DBBasicTestWithTimestamp, GetTimestampTableProperties) {
   ASSERT_EQ(2U, props.size());
   for (const auto& item : props) {
     auto& user_collected = item.second->user_collected_properties;
-    ASSERT_TRUE(user_collected.find("rocksdb.min_timestamp") !=
+    ASSERT_TRUE(user_collected.find("rocksdb.timestamp_min") !=
                 user_collected.end());
-    ASSERT_TRUE(user_collected.find("rocksdb.max_timestamp") !=
+    ASSERT_TRUE(user_collected.find("rocksdb.timestamp_max") !=
                 user_collected.end());
-    ASSERT_EQ(user_collected.at("rocksdb.min_timestamp"), Timestamp(0, 0));
-    ASSERT_EQ(user_collected.at("rocksdb.max_timestamp"), Timestamp(9, 0));
+    ASSERT_EQ(user_collected.at("rocksdb.timestamp_min"), Timestamp(0, 0));
+    ASSERT_EQ(user_collected.at("rocksdb.timestamp_max"), Timestamp(9, 0));
   }
   Close();
 }