@sunlike-Lipo 2021-11-11
@riversand963 
I fixed review comments and add a unitTest.  Could you please help me to review what else needs to be modified if that is convenient.