@sunlike-Lipo 2021-11-16
> @sunlike-Lipo Will take a look. One question: do you by any chance know the overhead introduced by tracking min/max timestamps for each file, since the comparison will be done for each key? It would be nice to include some benchmarking results here. We can assume timestamp size of 8 bytes.


I wrote a unittest(https://github.com/sunlike-Lipo/rocksdb/commit/d61cbab3dfbb3109378adca04dced8c0aaf24a3b) for the benchmark,
In general, the test runs for 10 times and collects the total time used for sstfiles' Flush.
I use the MockEnv to eliminate disk-IO's effect.

The variables are `value size` and `collect timestamp or not`,. The comparator used is ComparatorWithU64Ts.
the test results are list below

|                | value_size=10 | value_size=100 | value_size=1000 |
| -------------- | ------------- | -------------- | --------------- |
| collect-ts     | 4777ms        | 1517ms         | 644ms           |
| not-collect-ts | 4305ms        | 1458ms         | 619ms           |
| downgrade      | 10.9%         | 4.0%           | 4.0%            |