@riversand963 2021-11-16
Thanks @sunlike-Lipo for addressing the comments. Will take another look soon.

> I wrote a unittest

I wonder if we can move the code to a standalone source file (similar to examples/) and compile with rocksdb library built with `DEBUG_LEVEL=0` and see the performance difference in opt mode too. You can also use `db_bench`. In unit test, I think the DEBUG_LEVEl is at least 1, thus may already have some overhead, making the performance overhead of ts comparison less obvious.