@sunlike-Lipo 2021-11-17
> Thanks @sunlike-Lipo for addressing the comments. Will take another look soon.
> 
> > I wrote a unittest
> 
> I wonder if we can move the code to a standalone source file (similar to examples/) and compile with rocksdb library built with `DEBUG_LEVEL=0` and see the performance difference in opt mode too. You can also use `db_bench`. In unit test, I think the DEBUG_LEVEl is at least 1, thus may already have some overhead, making the performance overhead of ts comparison less obvious.

I fixed it to a stand alone test (https://github.com/sunlike-Lipo/rocksdb/commit/50a1152e1172b239036a096ddadc2d24ac338521) for the benchmark. The logical is not changed:
In general, the test runs for 10 times and collects the total time used for sstfiles' Flush.
I use the MemEnv to eliminate disk-IO's effect.

The variables are `value size` and `collect timestamp or not`,. The comparator still is ComparatorWithU64Ts.
the test results are list below

|                | value_size=10 | value_size=100 | value_size=1000 |
| -------------- | ------------- | -------------- | --------------- |
| collect-ts     | 4694ms        | 1486ms         | 652ms           |
| not-collect-ts | 4319ms        | 1464ms         | 634ms           |
| downgrade      | 8.6%          | 1.5%           | 2.8%            |

I will also test the db_bench, the result will be soon.