@sunlike-Lipo 2021-11-18
> Thanks @sunlike-Lipo for addressing the comments. Will take another look soon.
> 
> > I wrote a unittest
> 
> I wonder if we can move the code to a standalone source file (similar to examples/) and compile with rocksdb library built with `DEBUG_LEVEL=0` and see the performance difference in opt mode too. You can also use `db_bench`. In unit test, I think the DEBUG_LEVEl is at least 1, thus may already have some overhead, making the performance overhead of ts comparison less obvious.

db_bench test is here:
command:
```
$ TEST_TMPDIR=/tmp/rocksdb ./db_bench -benchmarks=filluniquerandom -level0_slowdown_writes_trigger 102400 -level0_stop_writes_trigger 102400  -disable_auto_compactions true -user_timestamp_size 8 -compression_type=none  -num=20000000
```
With ts collector       : 22.8MB/s, 22.5MB/s, 22.5MB/s, 22.6MB/s, 22.6MB/s
With out ts collector : 22.5MB/s, 22.3MB/s, 22.6MB/s, 22.7MB/s, 22.6MB/s

The detailed results are as follows:
With ts collector     : 
5.193 micros/op 192573 ops/sec 22.8  MB/S
5.263 micros/op 190004 ops/sec 22.5  MB/S  
5.266 micros/op 189885 ops/sec 22.5  MB/S
5.242 micros/op 190750 ops/sec 22.6  MB/S
5.228 micros/op 191274 ops/sec 22.6  MB/S
                                                                                                                                             
With out ts collector : 
5.256 micros/op 190257 ops/sec 22.5  MB/s
5.298 micros/op 188735 ops/sec 22.3 MB/s
5.229 micros/op 191234 ops/sec 22.6 MB/s
5.204 micros/op 192158 ops/sec 22.7 MB/s
5.233 micros/op 191098 ops/sec 22.6 MB/s