@sunlike-Lipo 2021-11-19
I made some tests again after changing string =  to string.assign

The db_bench result are as follows:
With ts collector :
5.175 micros/op 193250 ops/sec 22.9 MB/S
5.149 micros/op 194219 ops/sec 23.0 MB/S
5.223 micros/op 191443 ops/sec 22.6 MB/S
5.167 micros/op 193534 ops/sec 22.9 MB/S
5.150 micros/op 194183 ops/sec 23.0 MB/S

Without ts collector :
5.209 micros/op 191963 ops/sec 22.7 MB/s
5.201 micros/op 192256 ops/sec 22.7 MB/s
5.147 micros/op 194269 ops/sec 23.0 MB/s
5.230 micros/op 191211 ops/sec 22.6 MB/s
5.269 micros/op 189781 ops/sec 22.4 MB/s

The standalone test from https://github.com/sunlike-Lipo/rocksdb/commit/50a1152e1172b239036a096ddadc2d24ac338521 result are as follows:
use string.assign

|                | value_size=10 | value_size=100 | value_size=1000 |
| -------------- | ------------- | -------------- | --------------- |
| collect-ts     | 3851ms        | 1219ms         | 202ms           |
| not-collect-ts | 3745ms        | 1181ms         | 189ms           |

use string.operator=

|                | value_size=10 | value_size=100 | value_size=1000 |
| -------------- | ------------- | -------------- | --------------- |
| collect-ts     | 3868ms        | 1239ms         | 204ms           |
| not-collect-ts | 3767ms        | 1189ms         | 195ms           |

It seems there is little difference between the result using string.assign and using string.operator=, perhaps the compiler is 
clever enough to optmize this. 
Any way, I'll use string.assign because it ought to be faster than string.operator=.